New model creation logic
